@@ -2,7 +2,7 @@ package com.razi.furnitar;
 
 import android.app.Activity;
 import android.content.Intent;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 import android.view.View;
 
 import com.mikepenz.materialdrawer.AccountHeader;
