@@ -3,7 +3,7 @@ package com.razi.furnitar;
 import android.app.IntentService;
 import android.content.Intent;
 import android.os.Environment;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import java.io.File;
