@@ -3,11 +3,11 @@ package com.razi.furnitar;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.CountDownTimer;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.view.animation.LinearInterpolator;
-import android.widget.ImageView;
+import android.widget.TextView;
 
 public class splash_activity extends AppCompatActivity {
 
@@ -19,7 +19,7 @@ public class splash_activity extends AppCompatActivity {
         animation.setInterpolator(new LinearInterpolator());
         animation.setRepeatCount(Animation.INFINITE);
         animation.setDuration(700);
-        final ImageView splash = (ImageView) findViewById(R.id.splash);
+        final TextView splash = (TextView) findViewById(R.id.splash);
         splash.startAnimation(animation);
         new CountDownTimer(800, 100) {
             @Override
