@@ -1,219 +1,140 @@
 package com.razi.furnitar;
 
+
+import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
 
-import com.google.android.gms.auth.api.signin.GoogleSignIn;
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
-import com.google.android.gms.auth.api.signin.GoogleSignInClient;
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
-import com.google.android.gms.common.SignInButton;
-import com.google.android.gms.common.api.ApiException;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthCredential;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.auth.GoogleAuthProvider;
+import androidx.appcompat.app.AppCompatActivity;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import com.data.User;
+import com.network.Api;
+import com.network.AppConfig;
+import com.network.ServerResponse;
+import com.razi.furnitar.databinding.ActivityLoginBinding;
+import com.utils.Config;
+import com.utils.SharedPrefManager;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+import retrofit2.Retrofit;
 
 public class Login extends AppCompatActivity {
-    private static final int RC_SIGN_IN = 69;
-    SignInButton sBtn;
-    internetConnectivity it;
-    FirebaseAuth gAuth;
-    EditText userT, pass;
-    Button signin, signUp;
-    GoogleSignInClient mGoogleSignInClient;
-    FirebaseAuth.AuthStateListener aL;
-    protected void onDestroy() {
-        unregisterReceiver(it);
-        super.onDestroy();
-    }
+
+    private static final String TAG = "Login";
+    private final Context context = this;
+    private ActivityLoginBinding binding;
+    private String Email, Password;
+    private SharedPrefManager sharedPreferences;
+    private User user;
+    private String u_id, u_name, u_email, u_password, u_mob;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_login);
-        gAuth = FirebaseAuth.getInstance();
-        userT = findViewById(R.id.username);
-        pass = findViewById(R.id.password);
-        signin = findViewById(R.id.signin);
-        signUp = findViewById(R.id.signup_2);
-        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(getString(R.string.default_web_client_id))
-                .requestEmail()
-                .build();
-        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
-        sBtn = findViewById(R.id.sign_in_button);
-        sBtn.setOnClickListener(v -> {
-            switch (v.getId()) {
-                case R.id.sign_in_button:
-                    signIn();
-                    break;
-                // ...
+        binding = ActivityLoginBinding.inflate(getLayoutInflater());
+        View view = binding.getRoot();
+        setContentView(view);
+        sharedPreferences = new SharedPrefManager(context);
+        init();
+    }
+
+    private void init() {
+        binding.registerbtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Login.this.startActivity(new Intent(context, signUp.class));
+                Log.e(TAG, "onClick: Register Button");
             }
         });
-        signin.setOnClickListener(v -> signIn(userT.getText().toString(), pass.getText().toString()));
-        signUp.setOnClickListener(v -> startActivity(new Intent(Login.this, signUp.class)));
-        aL = firebaseAuth -> {
-            if (firebaseAuth.getCurrentUser() != null) {
-                FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-                String name = user.getDisplayName();
-                String email = user.getEmail();
-                Uri photoUrl = user.getPhotoUrl();
-                String uid = user.getUid();
-                common.currentUser = new user(name, uid, email, photoUrl);
-                startActivity(new Intent(Login.this, MainActivity.class));
-            }
-        };
-        IntentFilter in = new IntentFilter("android.net.conn.CONNECTIVITY_CHANGE");
-        it = new internetConnectivity();
-        registerReceiver(it, in);
-
+        binding.signinbtn.setOnClickListener(v -> doLogin());
     }
 
-    private void signIn() {
-        Intent signInIntent = mGoogleSignInClient.getSignInIntent();
-        startActivityForResult(signInIntent, RC_SIGN_IN);
+    private void doLogin() {
+        Email = binding.email.getText().toString();
+        Password = binding.password.getText().toString();
+        if (TextUtils.isEmpty(Email) && TextUtils.isEmpty(Password)) {
+            binding.email.setError("All fields are required");
+            binding.password.setError("All fields are required");
+        } else {
+            executeLogin(Email, Password);
+        }
     }
 
+    private void executeLogin(String email, String password) {
+        Log.e(TAG, "executeLogin: " + email);
+        Log.e(TAG, "executeLogin: " + password);
 
-    public void signIn(String email, String password) {
+        Retrofit retrofit = AppConfig.getRetrofit();
+        Api service = retrofit.create(Api.class);
 
-        String regExpn =
-                "^(([\\w-]+\\.)+[\\w-]+|([a-zA-Z]{1}|[\\w-]{2,}))@"
-                        + "((([0-1]?[0-9]{1,2}|25[0-5]|2[0-4][0-9])\\.([0-1]?"
-                        + "[0-9]{1,2}|25[0-5]|2[0-4][0-9])\\."
-                        + "([0-1]?[0-9]{1,2}|25[0-5]|2[0-4][0-9])\\.([0-1]?"
-                        + "[0-9]{1,2}|25[0-5]|2[0-4][0-9])){1}|"
-                        + "([a-zA-Z]+[\\w-]+\\.)+[a-zA-Z]{2,4})$";
+        Call<ServerResponse> call = service.login(email, password);
+        call.enqueue(new Callback<ServerResponse>() {
+            @Override
+            public void onResponse(Call<ServerResponse> call, Response<ServerResponse> response) {
+                if (response.body() != null) {
+                    ServerResponse serverResponse = response.body();
+                    if (!serverResponse.getError()) {
+                        Config.showToast(context, serverResponse.getMessage());
+                        sendUserData(serverResponse.getUser());
 
-        CharSequence inputStr = userT.getText().toString();
-
-        Pattern pattern = Pattern.compile(regExpn, Pattern.CASE_INSENSITIVE);
-        Matcher matcher = pattern.matcher(inputStr);
-
-        if (!matcher.matches()) {
-            userT.setError("Invalid Email");
-            return;
-        }
-
-
-        String passe = pass.getText().toString();
-        if (TextUtils.isEmpty(passe) || pass.length() < 6) {
-            pass.setError("Invalid Password");
-            return;
-        }
-
-        gAuth.signInWithEmailAndPassword(email, password)
-                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            // Sign in success, update UI with the signed-in user's information
-                            Log.d("OK", "signInWithEmail:success");
-                            FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-                            String name = user.getDisplayName();
-                            String email = user.getEmail();
-                            Uri photoUrl = user.getPhotoUrl();
-                            String uid = user.getUid();
-                            common.currentUser = new user(name, uid, email, photoUrl);
-                            startActivity(new Intent(Login.this, MainActivity.class));
-                            // FirebaseUser user = gAuth.getCurrentUser();
-                            //updateUI(user);
-                        } else {
-                            // If sign in fails, display a message to the user.
-                            Log.w("NOT OK", "signInWithEmail:failure", task.getException());
-                            Toast.makeText(Login.this, "Authentication failed.",
-                                    Toast.LENGTH_SHORT).show();
-                            //updateUI(null);
-                        }
-
-                        // ...
+                    } else {
+                        Config.showToast(context, serverResponse.getMessage());
                     }
-                });
+                } else {
+                    Config.showToast(context, response.body().getMessage());
+                }
+            }
+
+            @Override
+            public void onFailure(Call<ServerResponse> call, Throwable t) {
+                Config.showToast(context, t.getMessage());
+            }
+        });
+
+
     }
+
+    /*----------------------------------------------- Save User Data -----------------------------------------------*/
+
+    private void sendUserData(User user) {
+
+        SharedPrefManager sharedPrefManager = new SharedPrefManager(context);
+        sharedPrefManager.setInt("id", user.getUId());
+        sharedPrefManager.setString("name", user.getUName());
+
+        sharedPref();
+    }
+
 
     @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        // Result returned from launching the Intent from GoogleSignInApi.getSignInIntent(...);
-        if (requestCode == RC_SIGN_IN) {
-            Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(data);
-            handleSignInResult(task);
-            try {
-                // Google Sign In was successful, authenticate with Firebase
-                GoogleSignInAccount account = task.getResult(ApiException.class);
-                firebaseAuthWithGoogle(account);
-            } catch (ApiException e) {
-                // Google Sign In failed, update UI appropriately
-                Log.w("Yens", "Google sign in failed", e);
-                // ...
-            }
-        }
-    }
-
-    private void handleSignInResult(Task<GoogleSignInAccount> completedTask) {
-        try {
-            GoogleSignInAccount account = completedTask.getResult(ApiException.class);
-
-            // Signed in successfully, show authenticated UI.
-            //updateUI(account);
-        } catch (ApiException e) {
-            // The ApiException status code indicates the detailed failure reason.
-            // Please refer to the GoogleSignInStatusCodes class reference for more information.
-            Log.w("OK", "signInResult:failed code=" + e.getStatusCode());
-            // updateUI(null);
-        }
-    }
-
-    private void firebaseAuthWithGoogle(GoogleSignInAccount acct) {
-        Log.d("WOT", "firebaseAuthWithGoogle:" + acct.getId());
-
-        AuthCredential credential = GoogleAuthProvider.getCredential(acct.getIdToken(), null);
-        gAuth.signInWithCredential(credential)
-                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            // Sign in success, update UI with the signed-in user's information
-                            Log.d("One", "signInWithCredential:success");
-                            FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-                            String name = user.getDisplayName();
-                            String email = user.getEmail();
-                            Uri photoUrl = user.getPhotoUrl();
-                            String uid = user.getUid();
-                            common.currentUser = new user(name, uid, email, photoUrl);
-                            Log.i("Yes", common.currentUser.getName());
-                            //updateUI(user);
-                        } else {
-                            // If sign in fails, display a message to the user.
-                            Log.w("OK", "signInWithCredential:failure", task.getException());
-                            //updateUI(null);
-                        }
-
-                        // ...
-                    }
-                });
-    }
-
-    @Override
-    public void onStart() {
+    protected void onStart() {
         super.onStart();
-        gAuth.addAuthStateListener(aL);
+
+        sharedPref();
+
+    }
+
+    private void sharedPref() {
+        SharedPrefManager sharedPrefManager = new SharedPrefManager(context);
+        Config.user_id = sharedPrefManager.getInt("id");
+
+        if (Config.user_id != -1) {
+
+            openActivity(DashboardActivity.class);
+            finish();
+        }
+
+    }
+
+
+    private void openActivity(Class aclass) {
+        Intent intent = new Intent(context, aclass);
+        startActivity(intent);
     }
 }
